Add a toggle command.
    perl glue.pl toggle <direction> <device1> <device2> <default_device>

toggle will switch the audio between device1 and device2.  If the
audio isn't set to either of them, it will set it to the
default_device.  If there is no default_device, it will do nothing.

For example, toggle output and input between a headset and the
built-ins, defaulting to the built ins.

    perl glue.pl toggle output 'Plantronics 7xx' 'Built-In Output' 'Built-In Output'
    perl glue.pl toggle input  'Plantronics 7xx' 'Built-In Microphone' 'Built-In Microphone'

I don't use Alfred myself.  This code is the result of a CodeMentor.IO
session sponsored by Jeremy Biron.  I figured it's worth submiting
back to the project with Jeremy's permission.